+++ .github/workflows/PublishWorkbookWorkflow.yml
@@ -1,38 +1,44 @@
 name: PublishWorkbook
 on:
+  workflow_dispatch:
   pull_request:
     branches:
       - main
-    types: [closed]
+
 env:
   UNAME: ${{secrets.USERNAME}}
   PASS: ${{secrets.PASSWORD}}
   SERVURL: ${{secrets.SERVERURL}}
+
 jobs:
   Publish:
     runs-on: ubuntu-latest
+
     steps:
       - name: Checkout
         uses: actions/checkout@v2
         with:
           lfs: true
           fetch-depth: 0
+
       - name: Checkout LFS objects
         run: git lfs checkout
+
       - name: Setup Python
         uses: actions/setup-python@v3
         with:
           python-version: "3.10"
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip 
           pip install tableauserverclient==0.22
+
       - name: Publish Workbook
         run: |
-          ls
-          project_id=`jq '.workbooks' project-id.json`
-          if [ "$project_id" != '' ];then
-              python3 publish_workbook.py --server_url "${{env.SERVURL}}" --project_id "${project_id}" --password ${{env.PASS}} --username ${{env.UNAME}}
+          project_data=`jq '.workbooks' ./scripts/project-data.json`
+          if [ "$project_data" != '' ];then
+              python3 ./scripts/publish_workbook.py --username ${{env.UNAME}} --password ${{env.PASS}} --server_url "${{env.SERVURL}}" --project_data "${project_data}"
           else
-              echo "project id is not accessable."
+              echo "project data of json is not accessable."
           fi
