+++ README.md
@@ -1,1 +1,57 @@
-# TableawWorkbookCICD
+<!--
+The name of the project. Can also be a logo or ASCII Art Header (this [site](http://patorjk.com/software/taag/#p=display&f=Slant&t=Pro%20WebApp) can generate nice ASCII art)	
+-->	
+# Project Name
+
+<!---
+Some prose that describes the project, in as much detail as necessary to ensure the reader walks away with a basic understanding of the purpose of this repository. Diagrams included here can be useful to help the reader understand how the system is architectured. Links to other relevant documentation might also be useful.
+-->
+Clear description of what the project does. 
+
+<!-- 
+A description of how the code included in the project runs in various environments, including the URLs if that's appropriate
+-->
+## Where Does This Run
+
+<!--
+Step-by-step instructions on how to install the code and any necessary dependencies
+-->	
+## How to Use
+
+<!--
+Make liberal use of code blocks in the "How to Use" section, so it's easy to cut-and-paste the relevant commands into a terminal
+
+Example (Node):
+
+```bash
+# Clone the repository
+git clone git@github.com:invitation-homes/my-cool-repo.git
+
+# Install dependencies
+npm install
+
+# Run the application
+npm start
+```
+
+-->	
+```bash
+# Clone the repository
+git clone git@github.com:invitation-homes/my-cool-repo.git
+
+# Install dependencies
+# <how_to_install_dependencies>
+
+# Run the application
+# <how_to_start_the_application>
+```
+
+<!-- 
+Instructions on how to run locally -- necessary configuration files, secret configuration, etc. As a developer, this section combined with the previous section should give me all of the necessary information to check out and run the project locally	
+-->	
+### Local Development
+
+<!-- 
+How is the code built and deployed? Where does the project run? What linting and security checks are in place? Where can you view test results & static code analysis?
+-->	
+## CI/CD & Deployment
