--- project-id.json
@@ -1,14 +1,0 @@
-{
-  "workbooks": [
-    {
-      "name": "Film_Workbook",
-      "file_path": "film_workbook.twb",
-      "project_path": "Technology",
-      "option": {
-        "tags": ["sample", "temp"],
-        "hidden_views": ["Event by day"],
-        "show_tabs": true,
-        "description": "Sample description"
-      }
-    }     ]
-}
